requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
dlib==19.24.99
docopt==0.6.2
executing==1.2.0
fastjsonschema==2.19.1
filelock==3.9.0
fonttools==4.38.0
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
hub-sdk==0.0.3
hydra-core==1.3.1
idna==3.4
imageio==2.26.1
importlib-metadata==6.0.0
ipython==8.12.3
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
networkx==3.2.1
numpy==1.24.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.9.0.80
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.1.2
referencing==0.33.0
requests==2.28.2
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
scipy==1.10.0
seaborn==0.12.2
sentry-sdk==1.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.5
stack-data==0.6.2
supervision==0.3.2
sympy==1.12
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
tinycss2==1.2.1
torch==2.1.2
torchaudio==2.2.0+cpu
torchvision==0.16.2
tornado==6.4
tqdm==4.64.1
traitlets==5.8.1
triton==2.1.0
typing_extensions==4.4.0
ultralytics==8.0.239
Unidecode==1.3.8
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.2
wincertstore==0.2
yarg==0.1.9
zipp==3.11.0
